+++ obsiflask/pages/bookmarks.py
@@ -1,3 +1,6 @@
+"""
+Basic logic for operations with short links
+"""
 from threading import Lock
 import json
 
@@ -32,43 +35,69 @@ class NewLinkForm(FlaskForm):
 
 
 def load_links():
-    path = resolve_service_path(AppState.config.shortlink_path)
+    """
+    Loads links into AppState
+    """
+    if '{}' not in AppState.config.shortlink_path:
+        raise ValueError('shortlink_path must contain a {} placeholder')
     with _lock:
-        if path.exists():
-            with open(path) as inp:
-                AppState.shortlinks = json.loads(inp.read())
-        else:
-            logger.info('Could not find any shortlink')
-            for vault in AppState.config.vaults:
+        for vault in AppState.config.vaults:
+            path = resolve_service_path(
+                AppState.config.shortlink_path.format(vault))
+
+            if path.exists():
+                with open(path) as inp:
+                    AppState.shortlinks[vault] = json.loads(inp.read())
+            else:
+                logger.info(
+                    f'Could not find any shortlink for vault "{vault}"')
                 AppState.shortlinks[vault] = {}
 
 
-def add_and_save_links(vault: str, key: str, value: str | None):
-    path = resolve_service_path(AppState.config.shortlink_path)
+def change_and_save_links(vault: str, key: str, value: str | None = None):
+    """
+    Changes links and saves them
+
+    Args:
+        vault (str): vault name
+        key (str): key for a short link
+        value (str | None): url. If not set, will delete a value by key
+    """
+    if '{}' not in AppState.config.shortlink_path:
+        raise ValueError('shortlink_path must contain a {} placeholder')
+    path = resolve_service_path(AppState.config.shortlink_path.format(vault))
     with _lock:
         if value is not None:
             AppState.shortlinks[vault][key] = value
         else:
             del AppState.shortlinks[vault][key]
         with open(path, 'w') as out:
-            out.write(json.dumps(AppState.shortlinks))
+            out.write(json.dumps(AppState.shortlinks[vault]))
 
 
-def render_links(vault):
+def render_links(vault) -> str:
+    """
+    Rendering logic.
+    Processes form for creating link.
+    If gets 'key' in a get request, removes the short link
+
+    Returns:
+        str: rendered html
+    """
     form = NewLinkForm()
     alias = AppState.config.vaults[vault].short_alias or vault
     if form.validate_on_submit():
         if form.alias.data in AppState.shortlinks[vault]:
             flash(f'Link {form.alias.data} already registered', 'error')
         else:
-            add_and_save_links(vault, form.alias.data, form.link.data)
+            change_and_save_links(vault, form.alias.data, form.link.data)
             flash(f'Link "{form.alias.data}" added')
     elif 'link' in request.args:
         key = request.args['link']
         if key not in AppState.shortlinks[vault]:
             flash(f'Link "{key}" not found', 'error')
         else:
-            add_and_save_links(vault, key, None)
+            change_and_save_links(vault, key, None)
             flash(f'Link "{key}" deleted')
 
     return render_template('bookmarks.html',